backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pypdf==4.2.0
PyMuPDF
Pillow==10.3.0
google-generativeai==0.5.4
aiofiles==23.2.1
